Downgraded to MCreator 2021.3 (2022.1 kinda sucked)

+++ src/main/java/io/github/team_lodestar/transcendeum/procedures/BubbleClickEquipProcedure.java
@@ -62,7 +62,7 @@ public class BubbleClickEquipProcedure {
 				}
 				(((entity instanceof LivingEntity) ? ((LivingEntity) entity).getItemStackFromSlot(EquipmentSlotType.HEAD) : ItemStack.EMPTY))
 						.setCount((int) 1);
-				world.setBlockState(new BlockPos(x, y, z), Blocks.AIR.getDefaultState(), 3);
+				world.setBlockState(new BlockPos((int) x, (int) y, (int) z), Blocks.AIR.getDefaultState(), 3);
 			}
 		}
 	}
