+++ src/main/java/io/github/team_lodestar/transcendeum/procedures/ChrysaliumVineBlockValidPlacementConditionProcedure.java
@@ -37,8 +37,8 @@ public class ChrysaliumVineBlockValidPlacementConditionProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		return (world.getBlockState(new BlockPos(x, y + 1, z))).getBlock() == KelavusBlock.block
-				|| (world.getBlockState(new BlockPos(x, y + 1, z))).getBlock() == ChrysaliumVineBlock.block
-				|| (world.getBlockState(new BlockPos(x, y + 1, z))).getBlock() == ChrysaliumVineBottomBlock.block;
+		return (world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z))).getBlock() == KelavusBlock.block
+				|| (world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z))).getBlock() == ChrysaliumVineBlock.block
+				|| (world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z))).getBlock() == ChrysaliumVineBottomBlock.block;
 	}
 }
