+++ src/main/java/io/github/team_lodestar/transcendeum/procedures/ChrysaliumVineBottomNeighbourBlockChangesProcedure.java
@@ -38,9 +38,9 @@ public class ChrysaliumVineBottomNeighbourBlockChangesProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if ((world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == ChrysaliumVineBottomBlock.block) {
+		if ((world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == ChrysaliumVineBottomBlock.block) {
 			{
-				BlockPos _bp = new BlockPos(x, y, z);
+				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
 				BlockState _bs = ChrysaliumVineBlock.block.getDefaultState();
 				BlockState _bso = world.getBlockState(_bp);
 				for (Map.Entry<Property<?>, Comparable<?>> entry : _bso.getValues().entrySet()) {
