+++ src/main/java/io/github/team_lodestar/transcendeum/procedures/EfowWeedBreakChangeProcedure.java
@@ -39,10 +39,11 @@ public class EfowWeedBreakChangeProcedure {
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
 		if (!BlockTags.getCollection().getTagByID(new ResourceLocation("forge:efow_weed_growable"))
-				.contains((world.getBlockState(new BlockPos(x, y - 1, z))).getBlock())) {
+				.contains((world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock())) {
 			if (world instanceof World) {
-				Block.spawnDrops(world.getBlockState(new BlockPos(x, y, z)), (World) world, new BlockPos(x, y, z));
-				world.destroyBlock(new BlockPos(x, y, z), false);
+				Block.spawnDrops(world.getBlockState(new BlockPos((int) x, (int) y, (int) z)), (World) world,
+						new BlockPos((int) x, (int) y, (int) z));
+				world.destroyBlock(new BlockPos((int) x, (int) y, (int) z), false);
 			}
 		}
 	}
