+++ src/main/java/io/github/team_lodestar/transcendeum/procedures/ViriclingerBaseNeighbourBlockChangesProcedure.java
@@ -37,9 +37,9 @@ public class ViriclingerBaseNeighbourBlockChangesProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if (world.isAirBlock(new BlockPos(x, y - 1, z))) {
+		if (world.isAirBlock(new BlockPos((int) x, (int) (y - 1), (int) z))) {
 			{
-				BlockPos _bp = new BlockPos(x, y, z);
+				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
 				BlockState _bs = ViriclingerBlock.block.getDefaultState();
 				BlockState _bso = world.getBlockState(_bp);
 				for (Map.Entry<Property<?>, Comparable<?>> entry : _bso.getValues().entrySet()) {
